Log exceptions in Canner
Canner would log exceptions that occured in taggers, but
exceptions in Canner itself would only be printed to the screen.
They are now logged along with the values of local variables
in the stack.

Some errors still won't be logged because the logging system
itself has to be initialized and operational and the snapshot
directory has to exist and be writable.