Improve exception handling and exception logging.
If an exception is caught while a session is active, canner tries to
close the session in the exception handler.  An exception closing the
session would obscure the original error.  Now the original exception
is propagated.

Also, tweak the order and formatting of exception logging to ensure
that the log ends with a clear error message.