Add --interact option to allow user interaction with a device
The --interact option will connect to a device and script through the
login step, and then connects the terminal with the device.

Also added --rcfile and --oneshotrc options so passwords can be given
in a file instead of on the command line.  --oneshotrc is designed to
be used with a temporary file and automatically deletes it once it's
been used.