Find personalities using entry points.
Personalities no longer register themselves.  Instead, each
personality is listed as an entry point in setup.py (or in the
setup.py of another distribution thanks to setuptools).

Also, each personality now has a match() classmethod which is used to
determine if it is the right personality to use.  This is just the
first step towards the new "driver" model.